+++ appinfo.py
@@ -23,34 +23,37 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 # sometimes py2exe requires the following import to be commented
-#from __future__ import unicode_literals
+# from __future__ import unicode_literals
 
-#import builtins  # Python 3 compatibility
+# import builtins  # Python 3 compatibility
 import datetime as dt
-#import future  # Python 3 compatibility
-#import io  # Python 3 compatibility
+# import future  # Python 3 compatibility
+# import io  # Python 3 compatibility
 
 
 APP_NAME = 'easysetup'
-APP_VERSION = '0.0.9'
+APP_VERSION = '0.0.10'
 APP_LICENSE = 'GNU General Public License v2 or later (GPLv2+)'
-APP_URL = 'https://github.com/jcrmatos/easysetup'
 APP_AUTHOR = 'Joao Carlos Roseta Matos'
 APP_EMAIL = 'jcrmatos@gmail.com'
+APP_URL = 'https://github.com/jcrmatos/easysetup'
+APP_KEYWORDS = 'easy setup setuptools'
+
+# change classifiers to be correct for your application/module
 CLASSIFIERS = ['Development Status :: 4 - Beta',
                'Environment :: Console',
                'Intended Audience :: Developers',
-               'Natural Language :: English',
                'License :: OSI Approved ::' + ' ' + APP_LICENSE,
-               'Operating System :: OS Independent',
+               'Natural Language :: English',
+               'Operating System :: Microsoft :: Windows',
                'Programming Language :: Python',
+               'Programming Language :: Python :: 2',
                'Programming Language :: Python :: 2.7',
+               'Programming Language :: Python :: 3',
                'Programming Language :: Python :: 3.4',
-               'Topic :: Software Development :: Build Tools',
-              ]
-APP_KEYWORDS = 'easy setup setuptools'
+               'Topic :: Software Development :: Build Tools']
 
-COPYRIGHT = 'Copyright 2009-' + str(dt.date.today().year) + ' '  + APP_AUTHOR
+COPYRIGHT = 'Copyright 2009-' + str(dt.date.today().year) + ' ' + APP_AUTHOR
 
 APP_TYPE = 'application'  # it can be application or module