+++ easysetup/template/doc/conf.py
@@ -117,6 +117,10 @@ todo_include_todos = False
 # a list of builtin themes.
 html_theme = 'alabaster'
 
+on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
+if on_rtd:
+    html_theme = 'sphinx_rtd_theme'
+
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.