+++ easysetup/template/setup_utils.py
@@ -24,11 +24,15 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-#import builtins  # Python 3 compatibility
-#import future  # Python 3 compatibility
+# import builtins  # Python 3 compatibility
+import datetime as dt
+# import future  # Python 3 compatibility
+import glob
 import io  # Python 3 compatibility
 import os
+# import pprint as pp
 import sys
+# import sysconfig
 import time
 import zipfile as zip
 
@@ -36,6 +40,83 @@ import appinfo
 
 
 SYS_ENC = sys.getfilesystemencoding()
+
+
+def check_copyright():
+    """Check copyright on files that have to be updated manually."""
+    files = ['setup_utils.py', 'build.cmd', 'appinfo.py']
+    update_required = 0
+    for filename in files:
+        if os.path.isfile(filename):
+            with io.open(filename, encoding=SYS_ENC) as file_:
+                text = file_.readlines()
+            for line in text:
+                if appinfo.COPYRIGHT in line:
+                    break
+                if 'Copyright 2009-' in line:
+                    print('Copyright in ' + filename + ' is not updated.')
+                    update_required += 1
+                    break
+    if update_required:
+        sys.exit(1)
+
+
+def update_copyright():
+    """Update copyright on source and license files."""
+    files = glob.glob('*.py')
+    files = [file_ for file_ in files if file_ not in ['appinfo.py',
+                                                       'setup_utils.py']]
+    files += glob.glob(appinfo.APP_NAME + '/*.py')
+    for filename in files:
+        with io.open(filename, encoding=SYS_ENC) as file_:
+            text = file_.readlines()
+        new_text = ''
+        changed = False
+        for line in text:
+            if ((not changed) and (appinfo.COPYRIGHT not in line) and
+               ('Copyright 2009-' in line)):
+                new_text += '# ' + appinfo.COPYRIGHT + '\n'
+                changed = True
+            else:
+                new_text += line
+        if changed:
+            with io.open(filename, 'w', encoding=SYS_ENC) as file_:
+                file_.writelines(new_text)
+
+    filename = 'doc/conf.py'
+    if os.path.isfile(filename):
+        with io.open(filename, encoding=SYS_ENC) as file_:
+            text = file_.readlines()
+        new_text = ''
+        changed = False
+        doc_copyright = ("copyright = u'2009-" + str(dt.date.today().year) +
+                         ', ' + appinfo.APP_AUTHOR + "'")
+        for line in text:
+            if ((not changed) and ("copyright = u'2009-" in line) and
+               (doc_copyright not in line)):
+                new_text += doc_copyright + '\n'
+                changed = True
+            else:
+                new_text += line
+        if changed:
+            with io.open(filename, 'w', encoding=SYS_ENC) as file_:
+                file_.writelines(new_text)
+
+    filename = 'LICENSE.rst'
+    with io.open(filename, encoding=SYS_ENC) as file_:
+        text = file_.readlines()
+    new_text = ''
+    changed = False
+    for line in text:
+        if ((not changed) and (appinfo.COPYRIGHT not in line) and
+           ('Copyright ' + '2009-' in line)):
+            new_text += '        ' + appinfo.COPYRIGHT + '\n'
+            changed = True
+        else:
+            new_text += line
+    if changed:
+        with io.open(filename, 'w', encoding=SYS_ENC) as file_:
+            file_.writelines(new_text)
 
 
 def sleep(seconds=5):
@@ -124,7 +205,7 @@ def create_doc_zip():
     with zip.ZipFile('pythonhosted.org/doc.zip', 'w') as archive:
         for root, dirs, files in os.walk(doc_path):
             for file_ in files:
-                if not '.pdf' in file_:
+                if '.pdf' not in file_:
                     pathname = os.path.join(root, file_)
                     filename = pathname.replace(doc_path + os.sep, '')
                     archive.write(pathname, filename)
@@ -133,7 +214,8 @@ def create_doc_zip():
 def upd_usage_in_readme():
     """Update usage in README.rst."""
     if os.path.isfile(appinfo.APP_NAME + '/usage.txt'):
-        with io.open(appinfo.APP_NAME + '/usage.txt', encoding=SYS_ENC) as file_:
+        with io.open(appinfo.APP_NAME +
+                     '/usage.txt', encoding=SYS_ENC) as file_:
             usage_text = file_.read()
 
         with io.open('README.rst', encoding=SYS_ENC) as file_:
@@ -185,9 +267,70 @@ def change_sphinx_theme():
         if changed:
             with io.open('doc/conf.py', 'w', encoding=SYS_ENC) as file_:
                 file_.write(new_text)
-    except ImportError as error:
+    except ImportError:  # as error:
         pass
 
 
+# def std_lib_modules():
+#     """List all (not complete) Standard library modules."""
+#     std_lib_dir = sysconfig.get_config_vars('LIBDEST')[0]
+#     modules_lst = []
+#     for top, dirs, files in os.walk(std_lib_dir):
+#         for nm in files:
+#             if nm != '__init__.py' and nm[-3:] == '.py':
+#                 module = os.path.join(top, nm)[len(std_lib_dir)+1:-3].replace('\\','.')
+#                 if 'site-packages.' not in module:
+#                     modules_lst.append(os.path.join(top, nm)[len(std_lib_dir)+1:-3].replace('\\','.'))
+#     pp.pprint(modules_lst)
+
+
+# def non_std_lib_modules():
+#     """List all non Standard library modules."""
+#     site_lib_dir = sysconfig.get_config_vars('LIBDEST')[0]
+#     site_lib_dir += '/site-packages'
+#     modules_lst = []
+#     for top, dirs, files in os.walk(site_lib_dir):
+#         for nm in files:
+#             if nm != '__init__.py' and nm[-3:] == '.py':
+#                 modules_lst.append(os.path.join(top, nm)[len(site_lib_dir)+1:-3].replace('\\','.'))
+#     pp.pprint(modules_lst)
+
+
+# def docstr2readme():
+#     """Copy main module docstring to README.rst."""
+#     with io.open(appinfo.APP_NAME + '/' + appinfo.APP_NAME + '.py',
+#                  encoding=SYS_ENC) as file_:
+#         text = file_.readlines()
+#
+#     text2copy = appinfo.APP_NAME + '\n' + '=' * len(appinfo.APP_NAME) + '\n\n'
+#
+#     start_copy = False
+#     for line in text:
+#         if '"""' in line:
+#             if start_copy:
+#                 break
+#             else:
+#                 start_copy = True
+#         elif start_copy:
+#             text2copy += line
+#
+#     text2copy += '\n'
+#
+#     with io.open('README.rst', encoding=SYS_ENC) as file_:
+#         text = file_.readlines()
+#
+#     until_eof = False
+#
+#     for line in text:
+#         if 'Resources' in line or until_eof:
+#             text2copy += line
+#             until_eof = True
+#
+#     with io.open('README.rst', 'w', encoding=SYS_ENC) as file_:
+#         file_.writelines(text2copy)
+
+
 if __name__ == '__main__':
+    # import doctest
+    # doctest.testmod(verbose=True)
     eval(sys.argv[1])